C# Wrapper Implementation

+++ wrappers/dotnet/lib/Requests/CredentialRequestMetadata.cs
@@ -1,10 +1,10 @@
-using AnonCredsNet.Objects;
+using AnonCredsNet.Models;
 
 namespace AnonCredsNet.Requests;
 
 public class CredentialRequestMetadata : AnonCredsObject
 {
-    internal CredentialRequestMetadata(int handle)
+    internal CredentialRequestMetadata(long handle)
         : base(handle) { }
 
     internal static CredentialRequestMetadata FromJson(string json) =>
